# resources/swagger.yaml
openapi: 3.0.1
info:
  title: PotatoRSS
  description: 'Web RSS reader'
  version: 0.1.0
servers:
- url: http://localhost:8080/api/v1
tags:
- name: authorization
  description: Авторизация
- name: user
  description: Работа с пользователями
- name: source
  description: Работа с ресурсами
paths:
  /user:
    get:
      tags:
      - user
      summary: Получить пользователя
      responses:
        200:
          description: Успешно
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
      security:
        - bearerAuth: []
    put:
      tags:
        - user
      summary: Изменить профиль
      requestBody:
        description: Изменяемый пользователь
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
      responses:
        200:
          description: Успешно
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
      security:
        - bearerAuth: []
    delete:
      tags:
        - user
      summary: Удалить пользователя
      responses:
        200:
          description: successful operation
      security:
        - bearerAuth: []
  /user/registrate:
    post:
      tags:
        - user
      summary: Зарегестрировать пользователя
      requestBody:
        description: Новый пользователь
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
      responses:
        200:
          description: Успешно
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
  /user/auth:
    post:
      tags:
        - authorization
      summary: Авторизовать пользователя
      responses:
        200:
          description: Успешно
          content:
            text/html:
              schema:
                type: string
      security:
        - basicAuth: []
  /sources:
    get:
      tags:
        - source
      summary: Список ресурсов
      responses:
        200:
          description: Успешно
          content:
            application/json:
              schema:
                type: array
                items: 
                  $ref: '#/components/schemas/Source'
      security:
          - bearerAuth: []
    post:
      tags:
        - source
      summary: Добавить ресурс
      requestBody:
        description: Новый ресурс
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Source'
      responses:
        200:
          description: Успешно
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Source'
      security:
        - bearerAuth: []
  /sources/{sourceID}:
    get:
      tags:
        - source
      summary: Получить информацию о ресурсе
      parameters:
        - in: path
          name: sourceID
          schema:
            type: integer
          required: true
          description: Id ресурса
      responses:
        200:
          description: Успешно
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Source'
      security:
        - bearerAuth: []
    put:  
      tags:
        - source
      summary: Изменить ресурс
      parameters:
        - in: path
          name: sourceID
          schema:
            type: integer
          required: true
          description: Id ресурса
      requestBody:
        description: Изменяемый ресурс
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Source'
      responses:
        200:
          description: Успешно
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Source'
      security:
        - bearerAuth: []
    delete:
      tags:
        - source
      summary: Удалить ресурс
      parameters:
        - in: path
          name: sourceID
          schema:
            type: integer
          required: true
          description: Id ресурса
      responses:
        200:
          description: Успешно
      security:
        - bearerAuth: []
components:
  schemas:
    User:
      type: object
      properties:
        id:
          type: integer
          format: int64
        role_id:
          type: integer
          format: int64
        username:
          type: string
        password:
          type: string
        email:
          type: string
    Source:
      type: object
      properties:
        id:
          type: integer
          format: int64
        user_id:
          type: integer
          format: int64
        title:
          type: string
        description:
          type: string
        link:
          type: string
        
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
    basicAuth:
      type: http
      scheme: basic